修改删除方法的参数类型为 Long，并重命名相关方法以提高一致性和可读性

+++ BackendService/src/main/java/com/example/controller/CustomerController.java
@@ -75,7 +75,7 @@ public class CustomerController {
 
     // 删除顾客
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCustomer(@PathVariable int id) {
+    public ResponseEntity<Void> deleteCustomer(@PathVariable Long id) {
         boolean deleted = customerService.deleteCustomerById(id);
         return deleted ?
                 ResponseEntity.noContent().build() :
